+++ src/panels/config/energy/ha-config-energy.ts
@@ -2,10 +2,12 @@ import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import "../../../components/ha-svg-icon";
 import {
-  EnergyPreferences,
   EnergyPreferencesValidation,
-  getEnergyPreferences,
   getEnergyPreferenceValidation,
+  EnergyInfo,
+  EnergyPreferences,
+  getEnergyInfo,
+  getEnergyPreferences,
 } from "../../../data/energy";
 import "../../../layouts/hass-loading-screen";
 import "../../../layouts/hass-tabs-subpage";
@@ -36,6 +38,8 @@ class HaConfigEnergy extends LitElement {
   @property({ attribute: false }) public route!: Route;
 
   @state() private _searchParms = new URLSearchParams(window.location.search);
+
+  @state() private _info?: EnergyInfo;
 
   @state() private _preferences?: EnergyPreferences;
 
@@ -90,6 +94,7 @@ class HaConfigEnergy extends LitElement {
             .hass=${this.hass}
             .preferences=${this._preferences!}
             .validationResult=${this._validationResult!}
+            .info=${this._info}
             @value-changed=${this._prefsChanged}
           ></ha-energy-solar-settings>
           <ha-energy-battery-settings
@@ -115,7 +120,10 @@ class HaConfigEnergy extends LitElement {
   }
 
   private async _fetchConfig() {
+    this._error = undefined;
+
     const validationPromise = getEnergyPreferenceValidation(this.hass);
+    const energyInfoPromise = await getEnergyInfo(this.hass);
     try {
       this._preferences = await getEnergyPreferences(this.hass);
     } catch (e) {
@@ -130,6 +138,7 @@ class HaConfigEnergy extends LitElement {
     } catch (e) {
       this._error = e.message;
     }
+    this._info = await energyInfoPromise;
   }
 
   private async _prefsChanged(ev: CustomEvent) {
@@ -140,6 +149,7 @@ class HaConfigEnergy extends LitElement {
     } catch (e) {
       this._error = e.message;
     }
+    this._info = await getEnergyInfo(this.hass);
   }
 
   static get styles(): CSSResultGroup {
